requirements.txt
aiohttp==3.8.1
aiohttp-sse-client==0.2.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
dnspython==2.2.1
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
jsonpickle==2.2.0
Markdown==3.3.6
mnemonic==0.20
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pymongo==4.1.1
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
Random-Word==1.0.8
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stellar-base-sseclient==0.0.21
stellar-sdk==7.0.2
toml==0.10.2
tomli==2.0.1
typeguard==2.13.3
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
